requirements.txt
colorlog==6.7.0
easydict==1.10
matplotlib==3.5.3
numpy==1.21.5
opencv_contrib_python==4.6.0.66
pandas==1.3.5
Pillow==9.2.0
progress==1.6
pycocotools==2.0.5
PyYAML==6.0
scikit_image==0.19.3
scipy==1.7.3
tqdm==4.64.1
